Move class to platform/library_loader
