+++ todo.md
@@ -1,14 +1,15 @@
 ## TODO
+- Add logging
 - Add ImGui
 - Add Google Test
 - Render Command API (draw line, shapes etc. as data that doesn't depend on OpenGL)
-- Add TypeScript
+- Add TypeScript (maybe)
   - https://v8.dev/docs/embed
 
 ## Doing
-- Add DLL based hot reloading
 
 ## Done
+- Add DLL based hot reloading
 - Add OpenGL
 - Add build github action
 - Create SDL2 window
